Develop

See merge request s02-final/s02p31a104!45

+++ back/LAST/src/main/java/com/back/admin/service/BoardService.java
@@ -4,7 +4,6 @@ import com.back.admin.domain.board.Board;
 import com.back.admin.domain.board.BoardRepository;
 import com.back.admin.domain.experience.Experience;
 import com.back.admin.domain.experience.ExperienceRepository;
-import com.back.admin.web.dto.board.BoardResponseDto;
 import com.back.admin.web.dto.board.BoardSaveRequestDto;
 import com.back.admin.web.dto.board.BoardUpdateRequestDto;
 import lombok.RequiredArgsConstructor;
@@ -20,39 +19,62 @@ public class BoardService {
     private final BoardRepository boardRepository;
     private final ExperienceRepository experienceRepository;
 
-    // 모든 자소서 보여주기 selectAll
+
     public List<Board> selectAll() {
         return boardRepository.findAll();
     }
 
 
-    // 특정 학생의 자소서 보여주기 findBoardByStu_id
     @Transactional
-    public List<BoardResponseDto> findBoardByStu_no(Long stu_no) {
-        return boardRepository.findByStu_no(stu_no);
+    public Board findBoardByUser_no(Long user_no) {
+        return boardRepository.findByStu_no(user_no);
     }
 
 
-    // 자소서 저장 save
     @Transactional
-    public Long save(Long experience_no, BoardSaveRequestDto boardSaveRequestDto) {
+    public boolean save(Long experience_no, BoardSaveRequestDto boardSaveRequestDto) {
         Experience experience=experienceRepository.findByExperience_no(experience_no);
-        return boardRepository.save(boardSaveRequestDto.toEntity(experience)).getBoard_no();
-    }
-
-
-    // 자소서 수정 update
-    @Transactional
-    public boolean update(Long board_no, BoardUpdateRequestDto boardUpdateRequestDto) {
-        Board board = boardRepository.findByBoard_no(board_no);
-        board.update(boardUpdateRequestDto.getBoard_question(), boardUpdateRequestDto.getBoard_content());
+        boardRepository.save(boardSaveRequestDto.toEntity(experience));
         return true;
     }
 
-    // 자소서 삭제 delete
+
     @Transactional
-    public void delete(Long board_no){
+    public boolean update(Long board_no, Long user_no, BoardUpdateRequestDto boardUpdateRequestDto) {
         Board board = boardRepository.findByBoard_no(board_no);
-        boardRepository.delete(board);
+        Long board_user_id = board.getExperienceboard().getStudentexperience().getUser_no();
+        if (board_user_id.equals(user_no)) {
+            board.update(boardUpdateRequestDto.getBoard_question(), boardUpdateRequestDto.getBoard_content(),
+                    boardUpdateRequestDto.getBoard_company(),boardUpdateRequestDto.getBoard_keyword());
+            return true;
+        } else {
+            return false;
+        }
     }
+
+
+    @Transactional
+    public boolean delete(Long board_no, Long user_no){
+        Board board = boardRepository.findByBoard_no(board_no);
+        Long board_user = board.getExperienceboard().getStudentexperience().getUser_no();
+        if (board_user.equals(user_no)) {
+            boardRepository.delete(board);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+
+    @Transactional
+    public Board findByCompany(String board_company) {
+        return boardRepository.findByBoard_company(board_company);
+    }
+
+
+    @Transactional
+    public Board findByKeyword(String board_keyword) {
+        return boardRepository.findByBoard_keyword(board_keyword);
+    }
+
 }
