+++ back/LAST/src/main/java/com/back/admin/service/ExperienceService.java
@@ -2,6 +2,8 @@ package com.back.admin.service;
 
 import com.back.admin.domain.experience.Experience;
 import com.back.admin.domain.experience.ExperienceRepository;
+import com.back.admin.domain.user.User;
+import com.back.admin.domain.user.UserRepository;
 import com.back.admin.web.dto.experience.ExperienceResponseDto;
 import com.back.admin.web.dto.experience.ExperienceSaveRequestDto;
 import com.back.admin.web.dto.experience.ExperienceUpdateRequestDto;
@@ -14,46 +16,61 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class ExperienceService {
-    private final ExperienceRepository experienceRepository;
 
-    // 관리자가 모든 학생의 경험을 확인하기
+    private final ExperienceRepository experienceRepository;
+    private final UserRepository userRepository;
+
+
+    @Transactional
     public List<Experience> selectAll() {
         return experienceRepository.findAll();
     }
 
 
-    // 학생 개인의 경험 확인하기
     @Transactional
-    public List<ExperienceResponseDto> findExperienceByStu_id(Long stu_no) {
-        return experienceRepository.findByStu_no(stu_no);
-    }
-
-    // 개인의 경험을 연도별로 정리해주기
-
-
-    // 경험 저장
-    @Transactional
-    public Long save(Long experience_no, ExperienceSaveRequestDto experienceSaveRequestDto) {
-        Experience experience = experienceRepository.findByExperience_no(experience_no);
-        return experienceRepository.save(experienceSaveRequestDto.toEntity()).getExperience_no();
+    public Experience findByExperience(Long experience_no) {
+        return experienceRepository.findByExperience_no(experience_no);
     }
 
 
-    // 경험 수정
     @Transactional
-    public boolean update(Long experience_no, ExperienceUpdateRequestDto experienceUpdateRequestDto) {
-        Experience experience = experienceRepository.findByExperience_no(experience_no);
-        experience.update(experienceUpdateRequestDto.getExperience_start(), experienceUpdateRequestDto.getExperience_end(),
-                        experienceUpdateRequestDto.getExperience_title(), experienceUpdateRequestDto.getExperience_content());
-        return true;
+    public List<ExperienceResponseDto> findExperienceByStu_no(Long user_no) {
+        return experienceRepository.findByStu_no(user_no);
     }
 
 
-    // 경험 삭제
     @Transactional
-    public void delete(Long experience_no){
+    public void save(ExperienceSaveRequestDto experienceSaveRequestDto, Long user_no) {
+        User student = userRepository.findByUser_no(user_no);
+        experienceRepository.save(experienceSaveRequestDto.toEntity(student));
+    }
+
+
+    @Transactional
+    public boolean update(Long experience_no, String user_id_email, ExperienceUpdateRequestDto experienceUpdateRequestDto) {
         Experience experience = experienceRepository.findByExperience_no(experience_no);
-        experienceRepository.delete(experience);
+        String exp_user_id = experience.getStudentexperience().getUser_id_email();
+        if (exp_user_id.equals(user_id_email)) {
+            experience.update(experienceUpdateRequestDto.getExperience_start(), experienceUpdateRequestDto.getExperience_end(),
+                    experienceUpdateRequestDto.getExperience_title(), experienceUpdateRequestDto.getExperience_content(),
+                    experienceUpdateRequestDto.getExperience_tag());
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+
+    @Transactional
+    public boolean delete(Long experience_no, String user_id_email){
+        Experience experience = experienceRepository.findByExperience_no(experience_no);
+        String exp_user_id = experience.getStudentexperience().getUser_id_email();
+        if (exp_user_id.equals(user_id_email)) {
+            experienceRepository.delete(experience);
+            return true;
+        } else {
+            return false;
+        }
     }
 
 }
